refactor(api): extract modules from monolithic main.rs
Break up the 3078-line main.rs into focused modules:
- auth.rs: proof verification and signin flows (SSH + WebAuthn)
- schema.rs: GraphQL resolver types and enum conversions
- subscriptions.rs: subscription handlers and log streaming
- graphql_ws.rs: WebSocket connection state machine
- json_scalar.rs: JSON scalar conversion utilities

main.rs retains Query/Mutation roots, Context, CLI, and HTTP handlers,
dropping from 3078 to 994 lines.
